For btrfs, use btrfs send instead of the naive diff driver
Btrfs can quickly calculate the changes between two subvolumes, which is
significantly faster than using the naive diff driver (which involves
getting the stat for each file in each of the layers, as well as the
xattrs). Use this for determining the diff and pass in the changes list
to the archive functions.

Signed-off-by: Saikrishna Arcot <saiarcot895@gmail.com>